Modify the documentation to add link from original readme and fix the type of filesize

+++ src/main/java/com/woven/movie_uploader/mongo/FileMetadataModel.java
@@ -7,7 +7,7 @@ import org.springframework.data.annotation.Id;
 import java.util.Objects;
 
 public record FileMetadataModel(@Id String id, String name, String contentType,
-                                String createdAt, Binary content, int size) {
+                                String createdAt, Binary content, long size) {
     public FileMetadataModel {
         Objects.requireNonNull(id);
         Objects.requireNonNull(name);
@@ -18,7 +18,7 @@ public record FileMetadataModel(@Id String id, String name, String contentType,
 
     public FileMetadata convertToFileMetadata() {
         return new FileMetadata(
-                id,name, size, createdAt, contentType
+                id, name, size, createdAt, contentType
         );
     }
 }
